+++ project/runProjectTests.m
@@ -1,0 +1,1 @@
+testResults = runtests("../tests", IncludeSubfolders = true, ReportCoverageFor = ["../app", "../src"]);