+++ src/mission/bart/+mag/+bart/+view/Field.m
@@ -6,9 +6,8 @@ classdef Field < mag.graphics.view.View
         function this = Field(results, options)
 
             arguments
-                results
+                results (1, 1) mag.bart.Instrument
                 options.?mag.bart.view.Field
-                options.Results (1, 1) mag.bart.Instrument
             end
 
             this.Results = results;