+++ src/mission/imap/+mag/+imap/+view/Comparison.m
@@ -6,7 +6,7 @@ classdef Comparison < mag.graphics.view.View
         function this = Comparison(results, options)
 
             arguments
-                results
+                results (1, 1) mag.imap.Instrument
                 options.?mag.imap.view.Comparison
             end