+++ src/mission/imap/+mag/+imap/+view/EventPSD.m
@@ -13,7 +13,7 @@ classdef EventPSD < mag.graphics.view.View
         function this = EventPSD(results, options)
 
             arguments
-                results
+                results (1, 1) mag.imap.Instrument
                 options.?mag.imap.view.EventPSD
             end