+++ resources/release-notes.md
@@ -1,4 +1,17 @@
 # App
 
-- Reduce duplication of supported missions
-- Fix issue with opening app without mission
+- Add Signal Analyzer and Wavelet Analyzer as Bartington visualization options
+- Redesign mission selection window to support all enumerated missions (in `mag.meta.Mission`)
+- Redesign how view-controllers are selected for each mission
+- Reduce duplication in definition of Field, PSD and Spectrogram view-controllers
+
+# Software
+
+- Add support for dark-mode figures
+- Fix issue with tab-completion of Bartington views
+- `mag.Instrument.HasScience` now returns true if any science object has data
+
+# Project
+
+- Add shortcut to run tests with coverage
+- Reorganize project files